cryptobot_final/bot/requirements.txt
asyncpg==0.29.0
pybit==5.8.0
aiohttp==3.9.5
python-telegram-bot==21.5
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1
ta==0.11.0
scipy==1.13.1
